requirements.txt

streamlit
pandas
plotly
supabase
python-dotenv
prophet
cmdstanpy
requests
